Fixed LWA Formula
Fixed LWA formula - it now uses the M-Climatology and not the resulting standardized anomalies for gradient and mean. Also cleaned up the code a bit and patched a slight bug with M-Climate .npy files. Will need to fix M-Climates another time. Removed some test functions and started the process of getting rid of other bloat in the main M-Climate generating/LWA function.